split out the key based into a sperate component
it's made it's base component to hard to reuse and was only used once
